refactor: separate wrappers for wallet, connection, encryption states, export them from `react-ui`
